+++ encoding/json/inbound.go
@@ -43,7 +43,7 @@ type jsonHandler struct {
 	handler reflect.Value
 }
 
-func (h jsonHandler) Handle(ctx context.Context, _ transport.Options, treq *transport.Request, rw transport.ResponseWriter) error {
+func (h jsonHandler) Handle(ctx context.Context, treq *transport.Request, rw transport.ResponseWriter) error {
 	if err := encoding.Expect(treq, Encoding); err != nil {
 		return err
 	}
@@ -72,7 +72,7 @@ func (h jsonHandler) Handle(ctx context.Context, _ transport.Options, treq *tran
 	return nil
 }
 
-func (h jsonHandler) HandleOneway(ctx context.Context, _ transport.Options, treq *transport.Request) error {
+func (h jsonHandler) HandleOneway(ctx context.Context, treq *transport.Request) error {
 	if err := encoding.Expect(treq, Encoding); err != nil {
 		return err
 	}