+++ transport/handler.go
@@ -63,12 +63,7 @@ type Handler interface {
 	// An error may be returned in case of failures. BadRequestError must be
 	// returned for invalid requests. All other failures are treated as
 	// UnexpectedErrors.
-	Handle(
-		ctx context.Context,
-		opts Options,
-		req *Request,
-		resw ResponseWriter,
-	) error
+	Handle(ctx context.Context, req *Request, resw ResponseWriter) error
 }
 
 // OnewayHandler handles a single, transport-level, oneway request.
@@ -79,7 +74,6 @@ type OnewayHandler interface {
 	// TODO: determine oneway errors and how to deal with them
 	HandleOneway(
 		ctx context.Context,
-		opts Options,
 		req *Request,
 	) error
 }