+++ transport/http/inbound_test.go
@@ -140,7 +140,8 @@ func TestInboundMux(t *testing.T) {
 	spec := transport.HandlerSpec{Type: transport.Unary, Handler: h}
 	reg.EXPECT().GetHandlerSpec("bar", "hello").Return(spec, nil)
 
-	h.EXPECT().Handle(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
+	h.EXPECT().Handle(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
+
 	res, err := o.Call(ctx, &transport.Request{
 		Caller:    "foo",
 		Service:   "bar",